Add an add method to add together 2 vectors.
This is the first method where we really need to make sure that the
vectors' components match up nicely.  _checkVectorSpace was added in
order to help make this simple.

Because PHP arrays can have string keys, and because I'd like this class
to keep keys intact as much as possible, this _checkVectorSpace method
verifies the keys match between the arrays.  This means that trying to
add [1, 2] and ['x' => 1, 'y' => 2] together won't work - the keys don't
match.